Did some change on timestamp, change it type to Long and push to firestore

+++ app/src/main/java/fyp/ride_sharing_aos/model/Room.kt
@@ -1,5 +1,7 @@
 package fyp.ride_sharing_aos.model
 
+import java.sql.Timestamp
+
 
 /**
  * Created by lfreee on 21/1/2018.
@@ -12,8 +14,8 @@ class Room {
     var numOfPeople  : Int? = null
     var maleFil  : Boolean? = null
     var femaleFil   : Boolean? = null
-    var createtime  : Int? = null
-    var prefertime : Int? = null
+    var createtime  : Long? = null
+    var prefertime : Long? = null
     var roomname : String? = null
     var uid1 : String? = null
     var uid2 : String? = null
@@ -25,7 +27,7 @@ class Room {
         // Default constructor required for calls to DataSnapshot.getValue(User.class)
     }
 
-    constructor(start: String?, destination: String?, numOfPeople: Int?, maleFil: Boolean?, femaleFil: Boolean?, createtime: Int, prefertime: Int,roomname : String? , uid1: String, uid2: String, uid3:String, uid4: String) {
+    constructor(start: String?, destination: String?, numOfPeople: Int?, maleFil: Boolean?, femaleFil: Boolean?, createtime: Long, prefertime: Long,roomname : String? , uid1: String, uid2: String, uid3:String, uid4: String) {
         this.start = start
         this.destination = destination
         this.numOfPeople = numOfPeople
